+++ NEWS.md
@@ -1,3 +1,21 @@
+# lgr 0.4.4
+
+* `%k` and `%K` parameters in `format.LogEvent` now work as expected when using
+  **crayon** terminal colours (#52).
+  
+* Fix default format string for `LayoutGlue` which was using `msg` 
+  instead of `{msg}` (#54, thx @mmuurr)  
+  
+* Update docs to use the more common term "structured logging" instead of 
+  "custom fields" where appropriate
+
+* `as_event_list.data.frame` now really returns a list of `LogEvents`
+
+* added `as_LogEvent()` to coerce various event-like objects to `LogEvents`
+
+* rebuild docs for R 4.2.0
+
+
 # lgr 0.4.3
 
 * `logger_index()` returns a `data.frame` with metadata on all registered